Split CLI monolith into focused modules
Extract display helpers, command handlers, and utility functions from
the 2111-line main.rs into dedicated modules. main.rs now contains
only Clap struct definitions, arg parsing, and dispatch (~540 lines).

Closes #25
