+++ crates/pokeplanner-cli/src/commands/mod.rs
@@ -1,0 +1,6 @@
+pub mod cache;
+pub mod games;
+pub mod moves;
+pub mod pokemon;
+pub mod team;
+pub mod unusable_cmd;
